pyproject.toml
[project]
name = "ya-wiki-mcp"
version = "0.1.0"
description = "MCP server for Yandex Wiki API — pages, grids, YFM syntax, prompt templates"
requires-python = ">=3.11"
license = "MIT"
readme = "README.md"
keywords = ["mcp", "yandex", "wiki", "fastmcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
ya-wiki-mcp = "ya_wiki_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ya_wiki_mcp"]

[tool.hatch.build.targets.sdist]
include = ["src/ya_wiki_mcp"]
